% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{combine.experiments}
\alias{combine.experiments}
\title{Combine list of experiments into a matrix with the background model}
\usage{
combine.experiments(data, num.exp = 0, vp.pos)
}
\arguments{
\item{data}{list containing the 4C/CapC data in two column format}

\item{num.exp}{number of experiments, default is 0, which means the the number in the data list is taken, other values allow for choosing a subset of the experiments}

\item{vp.pos}{position of the viewpoint}
}
\value{
merged matrix containing: 1. the position of the fragments, 2:(n+1) the data, (n+1):(n+n+1) background models corresponding to the respective datasets
}
\description{
Combine list of experiments into a matrix with the background model
}
