% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{combined.analysis}
\alias{combined.analysis}
\title{Combined 4C/Capture C analysis}
\usage{
combined.analysis(data, num.exp = 0, vp.pos, wSize = 21, alphaFDR = 0.2,
  qW = 1, minDist = 15000)
}
\arguments{
\item{data}{list containing the 4C/CapC data in two column format, an additional element num.exp describes the number of experiments}

\item{num.exp}{number of experiments, used in conjuction with "data" and "multi" in type, default is 0 which means the number in the data list is used, a different number overwrites the default number}

\item{vp.pos}{viewpoint position, this can be a single value or a two values to analyse a viewpoint region}

\item{wSize}{number of fragments in a window}

\item{alphaFDR}{false-discovery rate threshold}

\item{qW}{threshold for absolute difference}

\item{minDist}{minimal region around the viewpoint to exclude for the significance analysis}
}
\value{
a list containing a matrix with the data and the background model and a vector with the significant fragments
}
\description{
Function for identifying interaction peaks above a background distribution. A list of 4C/Capture-C datasets are required as input. The viewpoint position is given in the vp.pos argument.
}
\examples{
data <- readMultiple(f[1:3], vp.pos = 65923803)
res <- combined.analysis(data, num.exp=3, vp.pos = 65923803)



}

