% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_functions.R
\name{data.frame.to.peakC}
\alias{data.frame.to.peakC}
\title{Transform a data frame to a list that can be used as input for peakC}
\usage{
data.frame.to.peakC(df, vp.pos, window, num.exp)
}
\arguments{
\item{df}{data.frame containing the experimental data}

\item{vp.pos}{position of the viewpoint}

\item{window}{flanking sequence to be considered in the analysis}

\item{num.exp}{Number of experiments}
}
\value{
A list containing two column matrices with the position and the normalized coverage score
}
\description{
Transform a data frame to a list that can be used as input for peakC
}
