% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{plot_C}
\alias{plot_C}
\title{Take a result from the combined.analysis function and generate a chromosomal map of the result}
\usage{
plot_C(data, num.exp = 0, y.min = 0, y.max = 3000, ...)
}
\arguments{
\item{data}{list containing the output of combined.analysis (i.e. 4C data and significant fragments)}

\item{num.exp}{number of experiments}

\item{y.min}{bottom limit of the plot}

\item{y.max}{top limit of the plot}
}
\value{
Nothing, a plot is drawn
}
\description{
Take a result from the combined.analysis function and generate a chromosomal map of the result
}
