% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_functions.R
\name{readMatrix}
\alias{readMatrix}
\title{Quickly read and normalize a matrix formatted file}
\usage{
readMatrix(file, window, vp.pos)
}
\arguments{
\item{file}{path to a two column file}

\item{window}{genomic window around the viewpoint to read in}

\item{vp.pos}{postion of the viewpoint in the genome}
}
\value{
a matrix with two columns, position and score
}
\description{
Wrapper function for reading files that are formatted as wig files. The data is also normalized to 1 million sequencing reads.
}
\examples{

}

