% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_functions.R
\name{readMultiColumnFile}
\alias{readMultiColumnFile}
\title{Read in a multiple experiment file}
\usage{
readMultiColumnFile(file, vp.pos, window = 7e+05, num.exp = 4)
}
\arguments{
\item{file}{path to the experiment file}

\item{vp.pos}{position of the viewpoint}

\item{window}{flanking sequence to be considered in the analysis}

\item{num.exp}{Number of experiments}
}
\value{
A list containing two column matrices with the position and the normalized coverage score
}
\description{
Read in a multiple experiment file
}

