% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{rem}
\alias{rem}
\title{Remove leading and trailing values}
\usage{
rem(a, n)
}
\arguments{
\item{a}{vector}

\item{n}{window size of the corresponding running function}
}
\value{
vector, shortened by (n - 1) elements
}
\description{
remove n/2 elements from the front and the end
}

