% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{running}
\alias{running}
\title{Title}
\usage{
running(x, n = 20)
}
\arguments{
\item{x}{numeric vector}

\item{n}{window size for the running window}
}
\value{
a numeric vector with the windowed means
}
\description{
Title
}

