% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{single.analysis}
\alias{single.analysis}
\title{Single experiment 4C/Capture C analysis}
\usage{
single.analysis(data, vp.pos, wSize = 21, qWd = 1.5, qWr = 1,
  minDist = 15000)
}
\arguments{
\item{data}{list containing the 4C/CapC data in two column format, an additional element num.exp describes the number of experiments}

\item{vp.pos}{viewpoint position, this can be a single value or a two values to analyse a viewpoint region}

\item{wSize}{number of fragments in a window}

\item{qWd}{threshold for difference from the background}

\item{qWr}{threshold for ratio over the background}

\item{minDist}{minimal region around the viewpoint to exclude for the significance analysis}

\item{alphaFDR}{false-discovery rate threshold}
}
\value{
a list containing a matrix with the data and the background model and a vector with the significant fragments
}
\description{
Function for identifying interaction peaks above a background distribution. A list containing a 4C/Capture-C dataset is required as input. The viewpoint position is given in the vp.pos argument.
}
\examples{
data <- readMultiple(f[1:3], vp.pos = 65923803)
res <- combined.analysis(data, num.exp=3, vp.pos = 65923803)



}
